requirements.txt
Telethon~=1.25.0
requests~=2.28.1
python-dateutil~=2.8.2